Fix building the ixgbevf module in the openSUSE 42.3 stemcell
openSUSE 42.3 kernels (4.4.x) include patches backported from newer versions. Some of
these patches are also part of the ixgbevf sources and applied depending on the kernel
version. These conditions do not work in this case because of the backports which leads
to conflicts. This patch should no longer be required once the openSUSE stemcell
ships a newer kernel (at least 4.8.x).

Signed-off-by: Dimitris Karakasilis <dkarakasilis@suse.com>